requirements.txt
flask>=2.3.3
flask-cors>=4.0.0
openai>=1.23.0
reportlab>=3.6.13
python-dotenv>=1.0.1
Flask-SQLAlchemy>=3.1